Use pyttsx instead of gtts and dump dict to json
Use pyttsx for text to speech implementation - it has less of a
lag and can read the text instead of strictly storing it to an
audio file. It removes the need to use playsound module as well.

Create a loop to iterate through the system-directed questions,
and collecting the user results.

Dump the dictionary of results to a JSON file.

Signed-off-by: Gabi Beyer <Gabib@live.com>